FSCORE-258 pcre compile warnings
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@10878 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/pcre/pcretest.c
@@ -794,7 +794,7 @@ offsets = (int *)malloc(size_offsets_max * sizeof(int));
 if (offsets == NULL)
   {
   printf("** Failed to get %d bytes of memory for offsets vector\n",
-    size_offsets_max * sizeof(int));
+    size_offsets_max * (int)sizeof(int));
   yield = 1;
   goto EXIT;
   }
@@ -1736,7 +1736,7 @@ while (!done)
           if (offsets == NULL)
             {
             printf("** Failed to get %d bytes of memory for offsets vector\n",
-              size_offsets_max * sizeof(int));
+              size_offsets_max * (int)sizeof(int));
             yield = 1;
             goto EXIT;
             }
